graph test 3, double fix

+++ Edge.java
@@ -1,10 +1,10 @@
 public class Edge
 {
     String to; // where does it go? immediate connection
-    int time; // how long will it take
+    double time; // how long will it take, double because decimals can't be read as int
     String line; // what colour line
 
-    public Edge(String to, int time, String line)
+    public Edge(String to, double time, String line)
     {
         this.to = to;
         this.time = time;
